Target elements are now focused by default
When an anchor pointing to an element is clicked, in addition to scroll to the target element, it will focus it.

This update fulfills the accessibility flaw when special users use the interface, those who rely on the TAB key to navigate in the page, like blind users or users with movement disability.

Note: In my opinion, for browser compatibility and user experience, you shouldn't use scrollIntoView(). For example, if you use only keyboard and click enter on the anchor, it will only scroll to the target element. If you make tab again, it will scroll all the way back where it was before when you clicked and not on the target element.

Please check this guideline and topics:
http://www.w3.org/TR/WCAG20-TECHS/SCR35.html
http://www.w3.org/TR/2009/WD-wai-aria-practices-20090224/#scrollintoview
http://www.quirksmode.org/dom/w3c_cssom.html